==> MachineLearning/ML_NLP_PySpark/Step1_Merging_2review_columns_into_1_column.ipynb <==
# importinng dependencies
import pandas as pd
import zipfile as zf

# import table that was cleaned earlier in Data_Preprocessing/Step1_Deleting_Null_Values_and_dividing_into_2_tables.ipynb
zip_file = zf.ZipFile('../../DataFiles/TransformedData/tables/hotel_reviews.zip') 
hotel_reviews = pd.read_csv(zip_file.open('hotel_review.csv'))
hotel_reviews

# Challenge 
* There are 16 columns to start with, however, we need only 2 for NLP - review and score
* there are 2 different columns for reviews: 'Positive reviews' and 'Negative Reviews'
* The column 'Positive reviews' has many entries as 'No Positive'
* The column 'Negative reviews' has many entries as 'No Negative'

# Solution: 
* Drop 'No Positive' and 'No negative' entries 
* Combining the 2 columns ('Positve Reviews' and 'Negative Reviews' into 1)
* Dropping all but 2 columns that are required for NLP

# changing the 'No Negative' entries to null
hotel_reviews.loc[(hotel_reviews["Negative_Review"] == "No Negative"), "Negative_Review"] = ''

# changing the 'No Positive' entries to null
hotel_reviews.loc[(hotel_reviews["Positive_Review"] == "No Positive"), "Positive_Review"] = ''

hotel_reviews

# merge positve and negative review columns into a 'Combined review' column
hotel_reviews["Review"] = hotel_reviews["Negative_Review"] + hotel_reviews["Positive_Review"]

hotel_reviews

# check for null values
hotel_reviews.isnull().sum()

# check for null values
hotel_reviews.isna().sum()

# Creating a table for PySpark NLP - 1st iteration

# retaining only two columns that are required for NLP - 'Review' and Reviewer_Score'
hotel_reviews = hotel_reviews[['Review','Reviewer_Score']]
hotel_reviews

# Exporting as a Zip file
compression_opts = dict(method='zip', archive_name='hotel_reviews_for_PySpark_preprocessing.csv')  
hotel_reviews.to_csv('PySpark_DataFile/hotel_reviews_for_PySpark_preprocessing.zip', index=False, compression=compression_opts)
