==> MachineLearning/ML_NLP_PySpark/Step2_Using_PySpark_to_clean_all_PySpark_Nulls.ipynb <==
# This file runs on ***Google Colab***
## Before running this file, the Data Source file needs to be placed at the same level as this file

### Data file name: PySpark_DataFile/hotel_reviews_for_PySpark_preprocessing.zip/hotel_reviews_for_PySpark_preprocessing.csv

import os
# Find the latest version of spark 3.0  from http://www-us.apache.org/dist/spark/ and enter as the spark version
# For example:
# spark_version = 'spark-3.0.1'
spark_version = 'spark-3.0.1'
os.environ['SPARK_VERSION']=spark_version

# Install Spark and Java
!apt-get update
!apt-get install openjdk-11-jdk-headless -qq > /dev/null
!wget -q http://www-us.apache.org/dist/spark/$SPARK_VERSION/$SPARK_VERSION-bin-hadoop2.7.tgz
!tar xf $SPARK_VERSION-bin-hadoop2.7.tgz
!pip install -q findspark

# Set Environment Variables
import os
os.environ["JAVA_HOME"] = "/usr/lib/jvm/java-11-openjdk-amd64"
os.environ["SPARK_HOME"] = f"/content/{spark_version}-bin-hadoop2.7"

# Start a SparkSession
import findspark
findspark.init()

# Start Spark session
from pyspark.sql import SparkSession
spark = SparkSession.builder.appName("Cleaning_Data_for_PySpark").getOrCreate()

# Read in data 
from pyspark import SparkFiles
file_path = "hotel_reviews_for_PySpark_preprocessing.csv"
spark.sparkContext.addFile(file_path)
df = spark.read.csv(SparkFiles.get("hotel_reviews_for_PySpark_preprocessing.csv"), sep=",", header=True)

# Show DataFrame
df.show()

# Checking for Null, nan and empty strings

from pyspark.sql.functions import isnan, when, count, col

# Checking the df for empty strings
df.filter(df["Review"] == "").count()

# Checking the df for null values
df.filter(df["Review"].isNull()).count()

# Checking the df for nan values
df.filter(isnan(df["Review"])).count()

# Retaining only NotNull values
df = df.where(col("Review").isNotNull())

# Checking again for any null values
df.filter(df["Review"].isNull()).count()

# Exporting cleaned data as CSV
df.toPandas().to_csv('data_ready_for_PySpark_NLP.csv', index=False)
