==> MachineLearning/ML_NLP_PySpark/Step4_NLP_Data_Preprocessing_for_Attempt2.ipynb <==
# importinng dependencies
import pandas as pd
import zipfile as zf
import numpy as np

# importing the Zip file 'CSV_cleaned_for_PySpark' and reading into a df
zip_file = zf.ZipFile('PySpark_DataFile/data_ready_for_PySpark_NLP.zip') 
hotel_reviews = pd.read_csv(zip_file.open('data_ready_for_PySpark_NLP.csv'))
hotel_reviews

# Challenge 
* Choosing the format of Reviewer_Score as float resulted in too big an output range for the PySpark Algorithm, which resulted in very low accuracy of the model: 20.77%

# Solution: 
* Rounding the format of Reviewer_Score to int

# creating a new column 'Reviewer_Score_int'
hotel_reviews['Reviewer_Score_Int'] = hotel_reviews['Reviewer_Score'].round(0).astype(int)
hotel_reviews

# drop the column 'Reviewer_Score'
hotel_reviews = hotel_reviews.drop(columns='Reviewer_Score')
hotel_reviews 


# Exporting as a Zip file
compression_opts = dict(method='zip', archive_name='hotel_reviews_for_NLP2_int_review_score.csv')  
hotel_reviews.to_csv('PySpark_DataFile/hotel_reviews_for_NLP2_int_review_score.zip', index=False, compression=compression_opts)
