==> MachineLearning/ML_NLP_PySpark/Step6_NLP_Data_Preprocessing_for_Attempt3.ipynb <==
# importinng dependencies
import pandas as pd
import zipfile as zf
import numpy as np

# importing the Zip file 'CSV_cleaned_for_PySpark' and reading into a df
zip_file = zf.ZipFile('PySpark_DataFile/hotel_reviews_for_NLP2_int_review_score.zip') 
hotel_reviews = pd.read_csv(zip_file.open('hotel_reviews_for_NLP2_int_review_score.csv'))
hotel_reviews

# Challenge 
* Choosing the format of Reviewer_Score as int resulted in an increase in accuracy, however the accuracy of 39.62% is still low

# Solution: 
* Convert the values in the column 'Reviewer_Score' to 5 star rating: 1 to 5

## Categories
* Star Rating 1: Integer Scores 0-2
* Star Rating 2: Integer Scores 3-4
* Star Rating 3: Integer Scores 5-6
* Star Rating 4: Integer Scores 7-8
* Star Rating 5: Integer Scores 9-10

# Create a new column 'Star_Rating'
hotel_reviews['Star_Rating'] = ''
hotel_reviews

# Converting values into Categories
hotel_reviews.loc[(hotel_reviews['Reviewer_Score_Int'] <= 2), 'Star_Rating'] = '1'
hotel_reviews.loc[(hotel_reviews['Reviewer_Score_Int'] >= 9), 'Star_Rating'] = '5'
hotel_reviews.loc[(hotel_reviews['Reviewer_Score_Int'] >= 3) & (hotel_reviews['Reviewer_Score_Int'] <= 4), 'Star_Rating'] = '2'
hotel_reviews.loc[(hotel_reviews['Reviewer_Score_Int'] >= 5) & (hotel_reviews['Reviewer_Score_Int'] <= 6), 'Star_Rating'] = '3'
hotel_reviews.loc[(hotel_reviews['Reviewer_Score_Int'] >= 7) & (hotel_reviews['Reviewer_Score_Int'] <= 8), 'Star_Rating'] = '4'
hotel_reviews

# Drop the column 'Reviewer_Score'
hotel_reviews = hotel_reviews.drop(columns ='Reviewer_Score_Int')
hotel_reviews

# Exporting as a Zip file
compression_opts = dict(method='zip', archive_name='hotel_reviews_for_NLP4_Star_Rating.csv')  
hotel_reviews.to_csv('PySpark_DataFile/hotel_reviews_for_NLP4_Star_Rating.zip', index=False, compression=compression_opts)
