==> MachineLearning/ML_NLP_PySpark/Step8_NLP_Data_Preprocessing_for_Attempt4.ipynb <==
# importinng dependencies
import pandas as pd
import zipfile as zf
import numpy as np

# importing the Zip file 'CSV_cleaned_for_PySpark' and reading into a df
zip_file = zf.ZipFile('PySpark_DataFile/hotel_reviews_for_NLP2_int_review_score.zip') 
hotel_reviews = pd.read_csv(zip_file.open('hotel_reviews_for_NLP2_int_review_score.csv'))
hotel_reviews

# Challenge 
* Choosing the format of Reviewer_Score as int resulted in an increase in accuracy, however the accuracy of 39.62% is still low

# Solution: 
* Convert the values in the column 'Reviewer_Score' to 3 categories - Bad, Average, and Good

## Categories
* Category 1: "Bad" - Score <= 4 
* Category 2: "Average" - Score 5-7
* Category 3: "Good" - Score >= 8

# Create a new column 'Review_Category'
hotel_reviews['Review_Category'] = ''
hotel_reviews

# Converting values into Categories
hotel_reviews.loc[(hotel_reviews['Reviewer_Score_Int'] <= 4), 'Review_Category'] = 'Bad'
hotel_reviews.loc[(hotel_reviews['Reviewer_Score_Int'] >= 8), 'Review_Category'] = 'Good'
hotel_reviews.loc[(hotel_reviews['Reviewer_Score_Int'] > 4) & (hotel_reviews['Reviewer_Score_Int'] < 8), 'Review_Category'] = 'Average'
hotel_reviews

# Drop the column 'Reviewer_Score'
hotel_reviews = hotel_reviews.drop(columns ='Reviewer_Score_Int')
hotel_reviews

# Exporting as a Zip file
compression_opts = dict(method='zip', archive_name='hotel_reviews_for_NLP3_3categories_review_score.csv')  
hotel_reviews.to_csv('PySpark_DataFile/hotel_reviews_for_NLP3_3categories_review_score.zip', index=False, compression=compression_opts)
