==> MachineLearning/ML_NLP_PySpark/Step10_NLP_Data_Preprocessing_for_Attempt5.ipynb <==
# importinng dependencies
import pandas as pd
import zipfile as zf

# importing the Zip file 'CSV_cleaned_for_PySpark' and reading into a df
zip_file = zf.ZipFile('PySpark_DataFile/data_ready_for_PySpark_NLP.zip') 
hotel_reviews = pd.read_csv(zip_file.open('data_ready_for_PySpark_NLP.csv'))
hotel_reviews

# Challenge 
* Choosing 3 categories improved the accuracy of our NLP algorithm is 77.63 %, now trying to imporve it further

# Solution: 
* Adding a new column 'Review_Sentiment' that will have the values as 'positive' and 'negative'
* Sentiment 'Positive' = Reviewer_Score >= 5
* Sentiment 'Negative' = Reviewer_Score <= 4.9

# creating a new column: 'Reviewer_Sentiment' - values 'positive' and 'negative'
hotel_reviews['Reviewer_Sentiment'] = ''
hotel_reviews

# For Scores <= 4.9, changing the value in 'review'

hotel_reviews.loc[hotel_reviews["Reviewer_Score"] <= 4.9, "Reviewer_Sentiment"] = "negative"
hotel_reviews.loc[hotel_reviews["Reviewer_Score"] >= 5.0, "Reviewer_Sentiment"] = "positive"

hotel_reviews

# dropping the Column 'Reviewer_Score'
hotel_reviews = hotel_reviews.drop(columns = 'Reviewer_Score')
hotel_reviews.index.name='Review_Id'
hotel_reviews

# Exporting as a Zip file
compression_opts = dict(method='zip', archive_name='hotel_reviews_for_NLP4_Positive_Negative_review_score.csv')  
hotel_reviews.to_csv('PySpark_DataFile/hotel_reviews_for_NLP4_Positive_Negative_review_score.zip', index=True, compression=compression_opts)
